Move overview tests which require running in serial
