+++ galata/test/documentation/overview.test.ts-snapshots/interface-jupyterlab-documentation-linux.png
