+++ galata/test/documentation/overview.test.ts-snapshots/interface-tabs-menu-documentation-linux.png
